Partially move text plugin stuff
